requirements.txt
imageio>=2.9.0
ipython>=7.29.0
matplotlib>=3.5.0
python>=3.7.0
pytorch>=1.10.0
scikit-image>=0.18.0
scipy>=1.7.1
tensorboard>=2.6.0
tensorflow>=2.4.1
torchvision>=0.11.1
